Update log config in `views.py`
Update the logging config in `views.py` to use the logging function
defined in `django_apps.utils.py` module.

The previous logging config was cluttering up the `views.py` module.
This ensures that the `views.py` will use the same logging config as the
rest of the file. Also brings the design in line with the
Dont Repeat Yourself (DRY) principle.