Refactor the Masker class
Refactor the masker class to remove the bug where it didn't mask the
video.

The bug was caused by the unintentional swapping of the upper and lower
ranges for each color range.

Also, there may have been other bugs, which were not investigated.
Instead we rewrote the `gen_video` method, and removed unnecessary
abstractions.

This way we avoided a costly and time consuming investigation.